feat - user 및 store 연결, 예외처리 변경

+++ src/main/java/org/sparta/foodordermanagementservice/service/ReviewServiceImpl.java
@@ -1,6 +1,8 @@
 package org.sparta.foodordermanagementservice.service;
 
 import lombok.RequiredArgsConstructor;
+import org.sparta.foodordermanagementservice.common.exeption.CustomException;
+import org.sparta.foodordermanagementservice.common.exeption.ErrorCode;
 import org.sparta.foodordermanagementservice.dto.request.ReviewListRequestDto;
 import org.sparta.foodordermanagementservice.dto.request.ReviewRequestDto;
 import org.sparta.foodordermanagementservice.dto.response.ReviewResponseDto;
@@ -21,18 +23,18 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class ReviewServiceImpl implements ReviewService {
 
-    private ReviewRepository reviewRepository;
-//    private StoreService storeService;
-//    private UserService userService;
+    private final ReviewRepository reviewRepository;
+    private final StoreService storeService;
+    private final UserService userService;
 
     // 리뷰 작성
     public void createReview(UUID storeId, ReviewRequestDto requestDto, String username) {
-//        Store store = storeService.findByStoreId(storeId);
-//        User user = userService.findByUsername(username);
+        Store store = storeService.findByStoreId(storeId);
+        User user = userService.findByUsername(username);
 
         Review review = Review.builder()
-//                .store(store)
-//                .user(user)
+                .store(store)
+                .user(user)
                 .rating(requestDto.getRating())
                 .content(requestDto.getContent())
                 .build();
@@ -44,10 +46,10 @@ public class ReviewServiceImpl implements ReviewService {
     @Transactional
     public void updateReview(UUID reviewId, ReviewRequestDto requestDto, String username) {
         Review review = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 리뷰를 찾을 수 없습니다."));
+                .orElseThrow(() -> new CustomException(ErrorCode.REVIEW_NOT_FOUND));
 
         if (!review.getUser().getUsername().equals(username)) {
-            throw new IllegalArgumentException("본인이 작성한 리뷰만 수정할 수 있습니다.");
+            throw new CustomException(ErrorCode.REVIEW_PERMISSION_DENIED);
         }
 
         review.update(requestDto);
@@ -56,11 +58,11 @@ public class ReviewServiceImpl implements ReviewService {
     // 리뷰 삭제
     public void deleteReview(UUID reviewId, String username) {
         Review review = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 리뷰를 찾을 수 없습니다."));
+                .orElseThrow(() -> new CustomException(ErrorCode.REVIEW_NOT_FOUND));
 
-//        if (!review.getUser().getUsername().equals(username) && !isMaster(username)) {
-//            throw new IllegalArgumentException("본인이 작성한 리뷰 또는 관리자만 삭제할 수 있습니다.");
-//        }
+        if (!review.getUser().getUsername().equals(username) && !isMaster(username)) {
+            throw new CustomException(ErrorCode.REVIEW_PERMISSION_DENIED);
+        }
 
         reviewRepository.delete(review);
     }
@@ -68,7 +70,7 @@ public class ReviewServiceImpl implements ReviewService {
     // 리뷰 단건 조회
     public ReviewResponseDto getReview(UUID reviewId) {
         Review review = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 리뷰를 찾을 수 없습니다."));
+                .orElseThrow(() -> new CustomException(ErrorCode.REVIEW_NOT_FOUND));
 
         return ReviewResponseDto.builder()
                 .rating(review.getRating())
@@ -79,7 +81,7 @@ public class ReviewServiceImpl implements ReviewService {
     // 가게 리뷰 목록 조회
     public Page<ReviewResponseDto> getReviewList(UUID storeId, ReviewListRequestDto requestDto) {
         Pageable pageable = PageRequest.of(requestDto.getPageNumber(), requestDto.getPageSize(),
-                requestDto.isAsc() ? Sort.by("created_at").ascending() : Sort.by("created_at").descending());
+                requestDto.isAsc() ? Sort.by("createdAt").ascending() : Sort.by("createdAt").descending());
 
 
         Page<Review> reviewPage = reviewRepository.findByStoreId(storeId, pageable);
@@ -95,10 +97,10 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     // 관리자 권한 체크 (ROLE_MASTER만 관리자)
-//    private boolean isMaster(String username) {x
-//        User user = userService.findByUsername(username);
-//        return user.getUserRole() == UserRole.MASTER;
-//    }
+    private boolean isMaster(String username) {
+        User user = userService.findByUsername(username);
+        return user.getUserRole() == UserRole.MASTER;
+    }
 
 
 }
